Moved ActivatedStep to setplib package
